+++ azimuth/modules/dataset_analysis/similarity_analysis.py
@@ -2,8 +2,8 @@
 # This source code is licensed under the Apache 2.0 license found in the LICENSE file
 # in the root directory of this source tree.
 import itertools
+import os
 from collections import defaultdict
-from os.path import join as pjoin
 from typing import Dict, List, Optional, Tuple
 
 import numpy as np
@@ -37,10 +37,17 @@ class FAISSModule(IndexableModule[SimilarityConfig]):
         self.encoder = None
         super().__init__(dataset_split_name, config, mod_options)
 
+    def get_model_name_or_path(self):
+        model_name_or_path = self.config.similarity.faiss_encoder
+        if os.environ.get("TRANSFORMERS_OFFLINE"):
+            home = os.environ.get("SENTENCE_TRANSFORMERS_HOME")
+            model_name_or_path = os.path.join(home, f"sentence-transformers_{model_name_or_path}")
+        return model_name_or_path
+
     def get_model(self):
         if self.encoder is None:
-            with FileLock(pjoin(self.cache_dir, "st.lock")):
-                self.encoder = SentenceTransformer(self.config.similarity.faiss_encoder)
+            with FileLock(os.path.join(self.cache_dir, "st.lock")):
+                self.encoder = SentenceTransformer(self.get_model_name_or_path())
         return self.encoder
 
     def compute_on_dataset_split(self) -> List[FAISSResponse]:  # type: ignore