+++ tests/test_routers/test_model_performance/test_outcome_count.py
@@ -24,3 +24,22 @@ def test_get_outcome_count_per_filter(app: FastAPI) -> None:
     data = resp.json()
     metrics = data.pop("countPerFilter")
     assert "label" in metrics and len(metrics["label"]) == 3
+
+
+def test_outcome_count_empty_filters(app: FastAPI):
+    client = TestClient(app)
+    resp = client.get(
+        "/dataset_splits/eval/outcome_count/per_filter?pipeline_index=0&utterance=yukongold"
+    )
+    assert resp.status_code == HTTP_200_OK, resp.text
+    data = resp.json()
+    assert data["utteranceCount"] == 0
+
+    # This used to fail when we were filtering on an empty dataset
+    resp = client.get(
+        "/dataset_splits/eval/outcome_count/per_filter"
+        "?pipeline_index=0&utterance=yukongold&data_action=relabel"
+    )
+    assert resp.status_code == HTTP_200_OK, resp.text
+    data = resp.json()
+    assert data["utteranceCount"] == 0