+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Released changes are shown in the
 
 ### Added
 - Added new class imbalance warnings
+- Added a new page to compare different pipelines in the performance analysis table.
 
 ### Changed
 - Added a proposed action: `merge_classes`, and rename `consider_new_class` to `define_new_class`.